Move Android plural resources code into new module
